+++ src/Stategies/AvalStrat/Question.py
@@ -1,6 +1,3 @@
-
-
-
 from typing import Union
 
 class Question:
